+++ src/main/java/org/styly/acanus/registry/ModEffects.java
@@ -13,6 +13,7 @@ import net.minecraftforge.registries.RegistryObject;
 import org.styly.acanus.Arcanus;
 import org.styly.acanus.effect.FullMoon;
 import org.styly.acanus.effect.MagicBlocked;
+import org.styly.acanus.effect.flight;
 
 public class ModEffects {
     public static final DeferredRegister<MobEffect> MOB_EFFECTS = DeferredRegister.create(ForgeRegistries.MOB_EFFECTS, Arcanus.MODID);
@@ -22,7 +23,7 @@ public class ModEffects {
     public static void register(IEventBus eventBus){
         MOB_EFFECTS.register(eventBus);
     }
-
+    public static final RegistryObject<MobEffect> FLIGHT = MOB_EFFECTS.register("flight",()->new flight(MobEffectCategory.BENEFICIAL,4393481));
     public static final RegistryObject<MobEffect> FULL_MOON = MOB_EFFECTS.register("full_moon",()->new FullMoon(MobEffectCategory.BENEFICIAL,3311322).addAttributeModifier(AttributeRegistry.SPELL_POWER.get(),"640667b9-5154-4547-84f7-a02a44f87782",0.50, AttributeModifier.Operation.MULTIPLY_TOTAL).addAttributeModifier(Attributes.ATTACK_DAMAGE,"340783c6-7234-481d-8ea5-68f39d82c15e",7, AttributeModifier.Operation.ADDITION).addAttributeModifier(AttributeRegistry.SPELL_RESIST.get(),"a0687470-3c16-45cf-8749-7966545f521c",0.5, AttributeModifier.Operation.MULTIPLY_TOTAL));
     public static final RegistryObject<MobEffect> MAGIC_BLOCKED = MOB_EFFECTS.register("magic_block",()->new MagicBlocked(MobEffectCategory.BENEFICIAL,4393481));
 }