+++ src/main/java/org/styly/acanus/registry/ModItems.java
@@ -15,6 +15,7 @@ import org.styly.acanus.Arcanus;
 import org.styly.acanus.curios.BaseArtifactItem;
 import org.styly.acanus.curios.BasicAttributeCurios;
 import org.styly.acanus.item.DarkMageArmourItem;
+import org.styly.acanus.item.FlightRing;
 
 public class ModItems {
 
@@ -35,6 +36,9 @@ public class ModItems {
     public static final RegistryObject<Item> WreathOfWinter = ITEMS.register("wreath_of_winter",()-> new BasicAttributeCurios(new Item.Properties().stacksTo(1).fireResistant(),AttributeRegistry.ICE_SPELL_POWER.get(), new AttributeModifier("ice",0.15, AttributeModifier.Operation.MULTIPLY_TOTAL)));
 
 
+    //Cool ring
+    public static final RegistryObject<Item> RING_OF_FLIGHT = ITEMS.register("flight_ring", FlightRing::new);
+    // Lore Armor
     public static final RegistryObject<Item> NETHERITE_MAGE_HELMET = ITEMS.register("netherite_mage_helmet", () -> new DarkMageArmourItem(ArmorItem.Type.HELMET, ItemPropertiesHelper.equipment()));
     public static final RegistryObject<Item> NETHERITE_MAGE_CHESTPLATE = ITEMS.register("netherite_mage_chestplate", () -> new DarkMageArmourItem(ArmorItem.Type.CHESTPLATE, ItemPropertiesHelper.equipment()));
     public static final RegistryObject<Item> NETHERITE_MAGE_LEGGINGS = ITEMS.register("netherite_mage_leggings", () -> new DarkMageArmourItem(ArmorItem.Type.LEGGINGS, ItemPropertiesHelper.equipment()));